requirement.txt
streamlit
openai
pdfminer.six
beautifulsoup4
requests
